Refactor ParallelMonitoring operation
